+++ netty/src/test/java/io/grpc/netty/NettyServerTest.java
@@ -63,7 +63,7 @@ public class NettyServerTest {
         new HashMap<ChannelOption<?>, Object>(),
         null, // no boss group
         null, // no event group
-        new ProtocolNegotiators.PlaintextNegotiator(),
+        ProtocolNegotiators.plaintext(),
         Collections.<ServerStreamTracer.Factory>emptyList(),
         TransportTracer.getDefaultFactory(),
         1, // ignore
@@ -101,7 +101,7 @@ public class NettyServerTest {
         new HashMap<ChannelOption<?>, Object>(),
         null, // no boss group
         null, // no event group
-        new ProtocolNegotiators.PlaintextNegotiator(),
+        ProtocolNegotiators.plaintext(),
         Collections.<ServerStreamTracer.Factory>emptyList(),
         TransportTracer.getDefaultFactory(),
         1, // ignore
@@ -139,7 +139,7 @@ public class NettyServerTest {
         channelOptions,
         null, // no boss group
         null, // no event group
-        new ProtocolNegotiators.PlaintextNegotiator(),
+        ProtocolNegotiators.plaintext(),
         Collections.<ServerStreamTracer.Factory>emptyList(),
         TransportTracer.getDefaultFactory(),
         1, // ignore
@@ -189,7 +189,7 @@ public class NettyServerTest {
         new HashMap<ChannelOption<?>, Object>(),
         null, // no boss group
         null, // no event group
-        new ProtocolNegotiators.PlaintextNegotiator(),
+        ProtocolNegotiators.plaintext(),
         Collections.<ServerStreamTracer.Factory>emptyList(),
         TransportTracer.getDefaultFactory(),
         1, // ignore
